Restructure constraints module
