+++ src/avh/constraints/_base.py
@@ -40,6 +40,9 @@ class Constraint(BaseEstimator):
         # self.differencing_lag = differencing_lag
         # self.preprocessing_func = preprocessing_func
 
+        self.differencing_lag = 0
+        self.preprocessing_func = utils.identity
+
     @classmethod
     def is_metric_compatable(self, metric: metrics.Metric):
         return issubclass(metric, self.compatable_metrics)
@@ -65,9 +68,9 @@ class Constraint(BaseEstimator):
 
     def fit(
         self,
-        column_history,
+        history: List[pd.Series],
         y=None,
-        # hotload_history: bool = False,
+        hotload_history: Optional[List[float]] = None,
         # preprocessed_metric_history: np.array = None,
         **kwargs,
     ) -> None:
@@ -76,7 +79,8 @@ class Constraint(BaseEstimator):
             f"{self.__class__.__name__}"
         )
 
-        self._fit(column_history, **kwargs)
+        self.metric_history_ = hotload_history if hotload_history is not None else self.metric.calculate(history)
+        self._fit(self.metric_history_, **kwargs)
 
         # self.metric_history_raw_ = (
         #     column_history if hotload_history else self.metric.calculate(column_history)
@@ -88,6 +92,12 @@ class Constraint(BaseEstimator):
         # )
 
         # self._fit(self.metric_history_post_, **kwargs)
+        return self
+    
+    def _fit(self, metric_history: List[float], **kwargs):
+        self.u_lower_ = 0.0
+        self.u_upper_ = 1.0
+        self.expected_fpr_ = 1.0
         return self
 
     def predict(self, column: pd.Series, **kwargs) -> bool:
@@ -102,12 +112,6 @@ class Constraint(BaseEstimator):
         # self.metric_history_raw_.append(metric_raw)
         # self.metric_history_post_.append(metric_post)
         return prediction
-
-    def _fit(self, metric_history: List[float], **kwargs):
-        self.u_lower_ = 0.0
-        self.u_upper_ = 1.0
-        self.expected_fpr_ = 1.0
-        return self
 
     def _predict(self, m: float, **kwargs) -> bool:
         return self.u_lower_ <= m <= self.u_upper_
@@ -126,95 +130,3 @@ class Constraint(BaseEstimator):
     #             data = diff(data, self.differencing_lag)
 
     #     return data
-
-
-class ConstantConstraint(Constraint):
-    """
-    Concrete Constraint subclass,
-        which operates on manually provided threshold values
-    """
-
-    def __init__(self, metric: metrics.Metric, u_lower: float, u_upper: float, expected_fpr):
-        super().__init__(metric)
-
-        # technically not following the sklearn style guide :(
-        self.u_upper_ = u_upper
-        self.u_lower_ = u_lower
-        self.expected_fpr_ = expected_fpr
-
-
-class ChebyshevConstraint(Constraint):
-    """
-    Chebyshev!
-    """
-
-    def _fit(self, metric_history: List[float], beta: float):
-        mean = np.nanmean(metric_history)
-        var = np.nanvar(metric_history)
-
-        self.u_upper_ = mean + beta
-        self.u_lower_ = mean - beta
-
-        if var == 0:
-            self.expected_fpr_ = 0.0
-        else:
-            self.expected_fpr_ = var / (beta**2)
-
-
-class CLTConstraint(Constraint):
-    compatable_metrics = (
-        metrics.RowCount,
-        metrics.Mean,
-        metrics.MeanStringLength,
-        metrics.MeanDigitLength,
-        metrics.MeanPunctuationLength,
-        metrics.CompleteRatio,
-    )
-
-    def _bell_function(sefl, x):
-        return math.pow(math.e, -(x**2))
-
-    def _fit(self, metric_history: List[float], beta: float):
-        mean = np.nanmean(metric_history)
-        std = np.nanstd(metric_history)
-
-        self.u_upper_ = mean + beta
-        self.u_lower_ = mean - beta
-
-        if std == 0:
-            self.expected_fpr_ = 0.0
-        else:
-            satisfaction_p = (2 / np.sqrt(math.pi)) * (
-                integrate.quad(self._bell_function, 0, beta / (np.sqrt(2) * std))[0]
-            )
-            self.expected_fpr_ = 1 - satisfaction_p
-
-
-class ConjuctivDQProgram:
-    def __init__(
-        self,
-        constraints: Optional[List[Constraint]] = None,
-        recall: Optional[Set[str]] = None,
-        contributions: Optional[List[Set[str]]] = None,
-    ):
-        self.constraints = constraints if constraints else []
-        self.recall = recall if recall else set({})
-        self.contributions = contributions if contributions else []
-
-    def __repr__(self):
-        return "{constraints}, FPR = {fpr:4f}".format(
-            constraints="\n".join([repr(q) for q in self.constraints]),
-            fpr=self.expected_fpr,
-        )
-
-    @property
-    def expected_fpr(self):
-        return sum([q.expected_fpr_ for q in self.constraints])
-
-    # def fit()
-
-    def predict(self, column: pd.Series, **kwargs):
-        for constraint in self.constraints:
-            if not constraint.predict(column, **kwargs):
-                return False
-        return True
